routing: refactoring

+++ modules/routing/include/motis/routing/output/transport.h
@@ -19,7 +19,7 @@ struct transport {
         mumo_price_(0) {}
 
   transport(unsigned const from, unsigned const to, unsigned const duration,
-            uint8_t const slot, unsigned const mumo_price)
+            unsigned const slot, unsigned const mumo_price)
       : from_(from),
         to_(to),
         con_(nullptr),
@@ -30,7 +30,7 @@ struct transport {
   unsigned from_, to_;
   light_connection const* con_;
   unsigned duration_;
-  uint8_t slot_;
+  unsigned slot_;
   unsigned mumo_price_;
 };
